added css and js files to example moved winkform source code to directory src to distuingish better between actual source code and the tests, example, doc and git files

+++ src/WinkForm/Input/Checkbox.php
@@ -4,12 +4,23 @@ class Checkbox extends Input
 {
     const ORIENTATION_HORIZONTAL = 0;
     const ORIENTATION_VERTICAL = 1;
-    
+
+    /**
+     * @var string
+     */
     protected $type = 'checkbox';
+
+    /**
+     * @var int
+     */
     protected $renderInColumns;
+
+    /**
+     * @var int
+     */
     protected $orientation = self::ORIENTATION_HORIZONTAL;
-    
-    
+
+
     /**
      * render the hidden input element
      */
@@ -22,7 +33,7 @@ class Checkbox extends Input
 
         // I will assume style adjustments apply to the container div if there are more
         // checkboxes and that it will apply to the checkbox if there is only one.
-                        
+
         // if it is a collection of checkboxes the property "values" is filled, otherwise the property "value"
         if (! empty($this->values))
         {
@@ -32,18 +43,18 @@ class Checkbox extends Input
         {
             $output .= $this->renderSingleCheckbox();
         }
-        
+
         // when posting, also send a hidden field so that if none of the options are selected
         // we still know that the checkbox has been posted
         $hidden = new HiddenInput($this->name.'-isPosted', 1);
         $output .= $hidden->render();
-        
+
         $output .= $this->renderInvalidations();
-        
+
         // return or echo output
         return $output;
     }
-    
+
     /**
      * render mutliple checkboxes
      * @return string $output
@@ -51,7 +62,7 @@ class Checkbox extends Input
     protected function renderMultipleCheckboxes()
     {
         $output = '';
-        
+
         // get the selected values
         $selected = $this->selected ?: $this->value;
         $selectedValues = array(); // default empty array
@@ -59,25 +70,25 @@ class Checkbox extends Input
         {
             $selectedValues = is_array($selected) ? $selected : array($selected);
         }
-        
+
         if (! empty($this->renderInColumns))
         {
             $rowsPerColumn = ceil(count($this->values) / $this->renderInColumns);
         }
-        
+
         $columns = array(); // array to collect the values per column in
         $classesAtStart = $this->classes;
         foreach ($this->values as $i => $value)
         {
             $checked = in_array($value, $selectedValues) ? ' checked="checked"' : '';
             $id = $this->id.'-'.$this->toValidHtmlId($value);
-        
+
             // add category class if it exists
             if (isset($this->categories[$i]))
             {
                 $this->addClass(str_replace(' ', '', $this->categories[$i]));
             }
-        
+
             $checkbox = '<input'
                 . $this->renderType()
                 . ' id="'.$id.'"'
@@ -90,12 +101,12 @@ class Checkbox extends Input
                 . $this->renderTitle()
                 . $this->renderDataAttributes()
                 . ' />'."\n";
-        
+
             if (isset($this->labels[$i])) // I also want to display 0
             {
                 $checkbox .= '<label for="'.$id.'">'.$this->labels[$i].'</label>'."\n";
             }
-        
+
             // render in columns per category, per given columns number, or don't render in columns
             if (! empty($this->categories[$i]))
             {
@@ -103,17 +114,17 @@ class Checkbox extends Input
             }
             elseif (! empty($this->renderInColumns))
             {
-                $columns[floor($i / $rowsPerColumn)][] = $checkbox;
+                $columns[(int) floor($i / $rowsPerColumn)][] = $checkbox;
             }
             else
             {
                 $output .= $checkbox;
             }
-        
+
             // reset classes, so the last entry doesn't have all category classes :)
             $this->classes = $classesAtStart;
         }
-        
+
         if (! empty($columns))
         {
             // recreate the $output
@@ -128,21 +139,21 @@ class Checkbox extends Input
                     $output .= '<input type="checkbox" class="'.$this->name.'-category" name="'.$categoryId.'" id="'.$categoryId.'">
                                     <label for="'.$categoryId.'" style="font-weight:bold;">'.$key."</label></strong><br/>\n";
                 }
-        
+
                 foreach ($col as $checkbox)
                 {
                     $output .= $checkbox."<br/>\n";
                 }
-        
+
                 $output .= '</td>';
-        
+
                 // go to new table row if the number of columns set in renderInColumns is reached. Add empty td for spacing.
                 if (! empty($this->categories) && ! empty($this->renderInColumns) && ++$colcount % $this->renderInColumns === 0)
                     $output .= '</tr><tr><td colspan="'.$this->renderInColumns.'" style="padding-top:10px;"></td></tr><tr>'."\n";
             }
-        
+
             $output .= '</tr></table>'."\n";
-        
+
             // add js for the category checkboxes
             if (! empty($this->categories))
             {
@@ -156,12 +167,12 @@ class Checkbox extends Input
                                 </script>'."\n";
             }
         }
-        
+
         $output = $this->renderLabel().'<div id="'.$this->id.'-container"'.$this->renderStyle().'>'."\n".$output."</div>\n";
-        
+
         return $output;
     }
-    
+
     /**
      * render single checkbox
      * @return string $output
@@ -174,9 +185,9 @@ class Checkbox extends Input
             if ($this->selected === true) // for intuitive use
                 $this->selected = '1';
         }
-        
+
         $checked = $this->value == $this->selected ? ' checked="checked" ' : '';
-        
+
         $checkbox = '<input'
             . $this->renderType()
             . $this->renderId()
@@ -190,15 +201,15 @@ class Checkbox extends Input
             . $this->renderDataAttributes()
             . $this->renderRequired()
             . ' />'."\n";
-        
+
         if (! empty($this->label))
         {
             $checkbox .= '<label for="'.$this->id.'">'.$this->label.'</label>'."\n";
         }
-        
+
         return '<div id="'.$this->id.'-container"'.$this->renderStyle().'>'."\n".$checkbox."</div>\n";
     }
-    
+
     /**
      * @param string $selected
      */
@@ -216,13 +227,13 @@ class Checkbox extends Input
                 else
                     $selected = array($selected);
             }
-                
+
             $this->selected = $selected;
         }
-        
+
         return $this;
     }
-    
+
     /**
      * store posted values
      * @param $posted
@@ -235,14 +246,14 @@ class Checkbox extends Input
             $posted = $_POST[$this->name];
             if (! empty($this->values) && ! is_array($posted))
                 $posted = array($posted);
-            
+
             $this->posted = $posted;
             $this->selected = $posted;  // so we can always retrieve the selected fields with getSelected()
         }
-        
+
         return $this;
     }
-    
+
     /**
      * is this input element posted?
      * @return boolean
@@ -252,7 +263,7 @@ class Checkbox extends Input
         // check the hidden input field that is always send
         return (isset($_POST[$this->name.'-isPosted']) && isPosted($this->name));
     }
-    
+
     /**
      * give amount of columns the list should be displayed in
      * @param int $int
@@ -263,10 +274,10 @@ class Checkbox extends Input
         {
             $this->renderInColumns = $int;
         }
-        
+
         return $this;
     }
-    
+
     /**
      * set the orientation (place the fields horizontally or vertically on screen)
      * @param string $orientation
@@ -279,24 +290,24 @@ class Checkbox extends Input
             $this->orientation = $orientation;
             $this->setRenderInColumns($orientation);
         }
-        
+
         return $this;
     }
-    
+
     /**
-     * @return renderInColumns
+     * @return int
      */
     public function getRenderInColumns()
     {
         return $this->renderInColumns;
     }
-    
+
     /**
-     * @return orientation
+     * @return int
      */
     public function getOrientation()
     {
         return $this->orientation;
     }
-    
+
 }
